requirements.txt
gradio
pytesseract
Pillow
transformers
torch
